fixed some button functionalities
