+++ src/components/ProfilePage/ProfileHeader.css
@@ -16,13 +16,10 @@
     width: 125px;
     height: 125px;
 	margin-left: 10px;
+	border-width: thick;
+	border-color: var(--background-color);
 }
-.picture-border{
-	flex: 1;
-	border-radius: 50%;
-	width: 100px;
-    height: 100px;
-}
+
 
 .info-container {
 	display: flex;
@@ -48,21 +45,26 @@
 .user-info{
 	flex-direction: column;
 	margin-top: 40px;
-	margin-right: 120px;
+	margin-right: 100px;
 	margin-bottom: 20px;
+	position: relative;
 }
 
 .headerOptions{
 	margin-top: 50px;
-	size: 3px;
+	margin-right: auto;
+	margin-left: auto;
+	size: 5px;
 	display: flex;
+	align-items: right;
 	gap: 3px;
+	padding-left: 5px;
 }
 
 .headerOptions button{
-	margin: 0px 1px;
+	margin: 0px 3px;
 	padding: 0px 7.5px;
-	height: 20%;
+	height: 25px;
 	background-color: var(--primary-color);
 	border-radius: 15px;
 	border-spacing: .5px;	
@@ -74,27 +76,3 @@
 	
 }
 
-.profileContentButtons{
-	display: flex;
-	gap: 5px;
-}
-
-.profileContentButtons button{
-	padding: 5px;
-	background-color: var(--background-color);
-	border-color: var(--background-color);
-	color: var(--text-color);
-	border-radius: 50px;
-	border-spacing: 10px;
-	border-width: 3px;
-}
-
-.profileContentButtons button:hover{
-	transition: background-color 250ms;
-	background: var(--primary-color);
-}
-
-.profileContentButtons button.active {
-	transition: background-color 250ms;
-	background-color: var(--primary-color);
-}
